// src/modules/auth/auth.interface.ts
export interface IRegisterPayload {
  name: string;
  email: string;
  password: string;
}

export interface ILoginPayload {
  email: string;
  password: string;
}

export interface IChangePasswordPayload {
  currentPassword: string;
  newPassword: string;
}

export interface IResetPasswordPayload {
  email: string;
  otp: string;
  newPassword: string;
}